insurance_backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-Mail==0.9.1
marshmallow==3.20.2
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
Pillow==10.2.0
python-magic==0.4.27
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
joblib==1.3.2
pandas==2.2.0
scikit-learn==1.4.0
pytest==8.0.0
black==24.1.1
flake8==7.0.0 